Started working on what I think the functions needed for queries will be

+++ db/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE department (
 CREATE TABLE role (
     id SERIAL PRIMARY KEY,
     title VARCHAR(30) UNIQUE NOT NULL,
-    salary INTEGER NOT NULL,
+    salary DECIMAL NOT NULL,
     department_id INTEGER NOT NULL,
     FOREIGN KEY (department_id)
     REFERENCES department(id)
